changed parseInt to parseFloat

+++ dps-ui/src/main/webapp/customers/controllers/addCustomerController.js
@@ -21,7 +21,7 @@ angular.module('addCustomerApp', ['ngMessages'])
 		    
 		    /* Function to calculate Additional Customer Margin */
 		    $scope.calculateCustomerMargin = function(){
-		    	$scope.customer.additionalMargin = parseInt($scope.customer.additionalMarginPercentage);
+		    	$scope.customer.additionalMargin = parseFloat($scope.customer.additionalMarginPercentage);
             	if($scope.customer.additionalMargin >= 0){
             		$scope.customer.additionalMargin = (1 / (1 - (Math.abs($scope.customer.additionalMargin)/100))).toFixed(3);
         		}
